requirements.txt
discord.py==1.6.0
dateparser==1.0.0
humanize==3.2.0